+++ src/components/editor/EditorCore.tsx
@@ -1,10 +1,14 @@
 import { useEditor, EditorContent } from "@tiptap/react"
 import { FloatingMenu, BubbleMenu } from "@tiptap/react/menus"
 import { extensions as baseExtensions } from "./extensions"
-import { useDebouncedCallback, getLoroExtensions } from "./hooks"
+import {
+  useDebouncedCallback,
+  getLoroExtensions,
+  useLoroDocForFile,
+} from "./hooks"
 import { updateFile } from "@/services/files"
 import { useCallback, useMemo } from "react"
-import { exportLoroSnapshotBase64, useLoroDocForFile } from "./loro"
+import { exportLoroSnapshotBase64 } from "./loro"
 
 export function EditorCore({
   fileId,
