+++ src/components/editor/hooks.ts
@@ -5,6 +5,7 @@ import { LoroSyncPlugin, LoroUndoPlugin } from "loro-prosemirror"
 import type { LoroDocType } from "loro-prosemirror"
 import { Extension } from "@tiptap/core"
 import { LoroDoc } from "loro-crdt"
+import { decodeBase64ToUint8 } from "./loro"
 
 export function useDebouncedCallback<Args extends unknown[]>(
   callback: (...args: Args) => void,
@@ -57,4 +58,48 @@ export function getLoroExtensions(loroDoc: LoroDoc) {
       ]
     },
   })
+}
+
+export function useLoroDocForFile(
+  fileId: string,
+  base64Content: string | null
+): { loroDoc: LoroDoc; markSnapshotApplied: (base64: string) => void } {
+  const loroDocRef = useRef<LoroDoc | null>(null)
+  const lastFileIdRef = useRef<string | null>(null)
+  const lastImportedBase64Ref = useRef<string | null>(null)
+
+  if (lastFileIdRef.current !== fileId || !loroDocRef.current) {
+    const doc = new LoroDoc()
+    if (base64Content) {
+      try {
+        const snapshot = decodeBase64ToUint8(base64Content)
+        doc.import(snapshot)
+        lastImportedBase64Ref.current = base64Content
+      } catch (e) {
+        console.warn("Loro import (sync) failed", e)
+      }
+    } else {
+      lastImportedBase64Ref.current = null
+    }
+    loroDocRef.current = doc
+    lastFileIdRef.current = fileId
+  }
+
+  useEffect(() => {
+    if (!loroDocRef.current || !base64Content) return
+    if (lastImportedBase64Ref.current === base64Content) return
+    try {
+      const snapshot = decodeBase64ToUint8(base64Content)
+      loroDocRef.current.import(snapshot)
+      lastImportedBase64Ref.current = base64Content
+    } catch (e) {
+      console.warn("Loro import failed", e)
+    }
+  }, [base64Content])
+
+  const markSnapshotApplied = (base64: string) => {
+    lastImportedBase64Ref.current = base64
+  }
+
+  return { loroDoc: loroDocRef.current!, markSnapshotApplied }
 }
